создает модуляцию
